refactor(edge): route /health through full worker pipeline
Replace main router shortcut with a proper health edge function at
functions/health/index.ts. Add PUBLIC_ROUTES set in main router to
bypass JWT for health checks while still exercising the full worker
spawn path — catches module loading, env var, and runtime issues.
