+++ apps/kbve/edge/functions/health/index.ts
@@ -1,0 +1,22 @@
+import { serve } from "https://deno.land/std@0.168.0/http/server.ts";
+import { corsHeaders } from "../_shared/cors.ts";
+
+const EDGE_VERSION = Deno.env.get("EDGE_VERSION") ?? "0.1.8";
+
+serve((req) => {
+  if (req.method === "OPTIONS") {
+    return new Response("ok", { headers: corsHeaders });
+  }
+
+  return new Response(
+    JSON.stringify({
+      status: "ok",
+      version: EDGE_VERSION,
+      timestamp: new Date().toISOString(),
+    }),
+    {
+      status: 200,
+      headers: { ...corsHeaders, "Content-Type": "application/json" },
+    },
+  );
+});
